requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis & Caching
redis==5.0.1
hiredis==2.2.3

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Data Processing
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2
xlrd==2.0.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Image Processing & OCR
Pillow==10.1.0
pytesseract==0.3.10

# Audio Processing
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1

# AI & ML
transformers==4.36.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.2

# Vector Database
chromadb==0.4.18
qdrant-client==1.7.0

# Language Models
langchain==0.0.350
langchain-community==0.0.10
langchain-experimental==0.0.47

# Fine-tuning
peft==0.7.1
accelerate==0.25.0
bitsandbytes==0.41.3

# TTS
TTS==0.22.0
phonemizer==3.2.1

# Monitoring & Logging
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.2

# Configuration
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Task Queue
celery==5.3.4
flower==2.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Deployment
gunicorn==21.2.0
docker==6.1.3

# Utilities
click==8.1.7
tqdm==4.66.1
python-slugify==8.0.1
jsonschema==4.20.0
marshmallow==3.20.1

# API Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Performance
asyncpg==0.29.0
ujson==5.8.0
orjson==3.9.10

# Security
cryptography==41.0.8
pyotp==2.9.0
qrcode==7.4.2

# File Processing
python-magic==0.4.27
chardet==5.2.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Environment Management
python-decouple==3.8